Alterações propostas na ultima aula 08/11/12
Criada classe Agência;
Alterada a forma de relacionamento entre as contas e as operações
conforme proposto pelo Bira na aula;
Aplicado padrão singleton nos repositórios;

+++ Infnet.Proj.SistBancario/Infnet.Proj.SistBancario.Modelo/PessoaFisica.cs
@@ -7,13 +7,13 @@ namespace SistBancario.Modelo
 {
     public class PessoaFisica : Cliente 
     {
-        public PessoaFisica(string Nome,decimal renda,int cpf):base(Nome)
+        public PessoaFisica(string Nome,double renda,int cpf):base(Nome)
         {
             this.Renda = renda;
             this.CPF = cpf;
         }
 
-        public decimal Renda { get; set; }
+        public double Renda { get; set; }
         public int CPF { get; set; } 
     }
 }
