+++ Infnet.Proj.SistBancario/Infnet.Proj.SistBancario.Modelo/PessoaJuridica.cs
@@ -7,13 +7,13 @@ namespace SistBancario.Modelo
 {
     public class PessoaJuridica : Cliente
     {
-        public PessoaJuridica(string Nome, decimal receita,int cnpj):base(Nome)
+        public PessoaJuridica(string Nome, double receita,int cnpj):base(Nome)
         {
             this.Receita = receita;
             this.CNPJ = cnpj;
         }
 
-        public decimal Receita { get; set; }
+        public double Receita { get; set; }
         public int CNPJ { get; set; } 
     }
 }
